loading refactor
