+++ packages/ui/src/viewer-table/Loading/index.ts
@@ -21,5 +21,4 @@
 
 export { HeaderSkeleton } from './HeaderSkeleton';
 export { InteractionPanelSkeleton } from './InteractionPanelSkeleton';
-export { default as LoadingPage } from './LoadingPage';
 export { default as LoadingSpinner } from './LoadingSpinner';
