+++ packages/ui/stories/viewer-table/DictionaryHeader.stories.tsx
@@ -25,7 +25,7 @@ import type { Meta, StoryObj } from '@storybook/react';
 import { pick } from 'lodash';
 
 import DictionaryHeader from '../../src/viewer-table/DictionaryHeader';
-import { HeaderSkeleton } from '../../src/viewer-table/Loading/HeaderSkeleton';
+import { HeaderSkeleton } from '../../src/viewer-table/Loading';
 import biosampleDictionary from '../fixtures/minimalBiosampleModel';
 import themeDecorator from '../themeDecorator';
